requirements.txt
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-debug-toolbar==1.9.1
django-extra-views==0.6.4
django-haystack==2.6.1
django-oscar==1.5
django-phonenumber-field==1.3.0
django-tables2==1.16.0
django-treebeard==4.1.2
django-widget-tweaks==1.4.1
factory-boy==2.9.2
Faker==0.8.7
httplib2==0.10.3
idna==2.6
mock==2.0.0
oauth2==1.9.0.post1
oauthlib==2.0.6
olefile==0.44
pbr==3.1.1
phonenumberslite==8.8.6
Pillow==4.3.0
purl==1.3.1
PyJWT==1.5.3
python-dateutil==2.6.1
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
text-unidecode==1.1
Unidecode==0.4.21
urllib3==1.22
